% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penguin_plot_function.R
\name{plot_species_distribution}
\alias{plot_species_distribution}
\title{Plot penguin variable}
\usage{
plot_species_distribution(data, variable_name)
}
\arguments{
\item{data}{Loads data set of penguin statistics}

\item{variable_name}{Select which variable you want summarized}
}
\value{
Returns a histogram (ggplot) of each species based on a specified numeric variable
}
\description{
Plots a histogram of the selected attribute for each of the palmerpenguins species.
}
\examples{
Plot the distribution of 'body_mass_g' for each penguin species
plot_species_distribution(palmerpenguins::penguins, "body_mass_g")
}
