% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penguin_summary_function.R
\name{summarize_species}
\alias{summarize_species}
\title{Summarize penguin data}
\usage{
summarize_species(data, variable)
}
\arguments{
\item{data}{Loads data set of penguin statistics}

\item{variable}{Select which variable you want summarized}
}
\value{
Returns a dataframe of the different aspects of selected variables
}
\description{
Returns a data frame with summary statistics. Useful for quickly summarising penguins by attributes like flipper length or body mass.
}
\examples{
summarize_species(data = palmerpenguins::penguins,
                 variable = flipper_length_mm)
}
